Opengl refactor

+++ src/engine/utils/Math.h
@@ -92,7 +92,7 @@ inline float magnitude2(const vec2& vec) {
 }
 
 inline float magnitude(const vec2& vec) {
-    return std::sqrtf(magnitude2(vec));
+    return std::sqrt(magnitude2(vec));
 }
 
 // random numbers
